Tested Solution Detection and Recorder
Added some test audio files to check if all this is really working. It appears like it is.

Right now, pressing K will enter your current setup for analysis and play the corresponding composition piece. (This will later change to an OnTriggerExit at the end of the level)

Currently, if no solution can be found, the first solution in a Song's possibleSolutions array will be played as the defult.