Sunbeams Initial Version Complete.
Initial prototype of sunbeams is working - but currently suffers from mesh artfacts when trying to render the beam.

Note that the collider works perfectly.

To repair this, the starting points of our ray checks need to be stored in a larger meshPoint array - and out collider must be constructed from a smaller colliderPoints array.

This way, each raycast can be treated as a thin, long recangle, which can be split into 2 triangles relatively easily.

I'll do that tomorrow though.