Completed Sunbeams
Sunbeams now move between a set of designated points in a cyclical order - moving to the next point each time their OnActiveBeat function is called.

They construct a collider along their edges which the player can stand on and will be blocked by.

They also construct a dynamic mesh which can be given a sprite-type material - allowing us to customise their appearance using shaders.
This is an optional function, and can be turned off in favour of using Unity's lighting system for representation.