Completed Sunbeams Script Added Into Project.
Dumb workaround to repair a merge error of my causing - apologies for the inconvinience.

The sunbeams are now fuly finctional - they can be given a set of transforms to move between, and will lerp to the next point in this array each time their OnActiveBeat event is called.

They construct a dynamic polygon collider around their edge - allowing us to block the player off or use them as platforms as discussed.

They also create a dynamic mesh to visualise themselves - this mesh can be assigned any material that would work in a sprite Renderer - thus we can customise their look with shaders.
This is optional, we can switch it off and design their look with Unity lighting too.