Started 3D Sunbeams. improved Solution Detection.
3D sunbeams (Collider approximation version) is essentially complete - but as we've decided to potentially scrap the idea, it's development has been put on hold.

The solution dectection script has been improved to check if the player has found a viable solution that's been shifted by a certain number of beats. Essentially, a user can enter a beat pattern that matches that in one of our pre-defined solutions, but it does not start at beat 0. The detector now picks this up. The audio played in response should be delayed by the correct number of beats now, but this still needs to be tested.