Tested Final COmposition.
When a solution is found, a composition matching it is played in response. After some testing, this system is functioning with some minor irregularities due to the BeatSoundPlayers not playing a single sound, but rather a selected note from a collection of sounds.

We'll discuss how to address this - either by randomising this note or by sticking to a single note for each BeatSoundPlayer.